docker/snmp-collector/requirements.txt
pysnmp-lextudio==5.0.33
